Fixed all memory crashes
Changed destroying functions for all managers
Added isDead bool to every object that was going to be delated
Replaced every Destroyer call by a clear.
Objects are now flagged as dead and destroyed in a cleaner way after the loop (and not into it).
Fixed about 10 warnings.

+++ Zenirox/globalvar.hpp
@@ -16,8 +16,8 @@ enum ID {
 
 
 
-const float HEIGHT = 1080;
-const float WIDTH = 1920;
-const float MAXHEIGHT = 95;
+const unsigned int HEIGHT = 1080;
+const unsigned int WIDTH = 1920;
+const unsigned int MAXHEIGHT = 95;
 
 #endif
